+++ pulsar/internal/blocking_queue.go
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package util
+package internal
 
 import (
 	"sync"
